ImageInfo: only call System.exit in batch mode
The ability to run showinf in batch mode introduced in Bio-Formats 8
has caused some regression with the interactivity of the utility when
opening a single image without the -nopix option as the System.exit
call immediately closes the viewer. This commit addresses the issue
by only invoking it when multiple input are passed.